+++ src/poetry/core/_vendor/lark/parsers/earley_forest.py
@@ -4,7 +4,7 @@ This is used as the primary output mechanism for the Earley parser
 in order to store complex ambiguities.
 
 Full reference and more details is here:
-http://www.bramvandersanden.com/post/2014/06/shared-packed-parse-forest/
+https://web.archive.org/web/20190616123959/http://www.bramvandersanden.com/post/2014/06/shared-packed-parse-forest/
 """
 
 from random import randint
@@ -770,7 +770,7 @@ class ForestToPyDotVisitor(ForestVisitor):
         graph_node = self.graph.get_node(graph_node_id)[0]
         for child in [node.left, node.right]:
             if child is not None:
-                child_graph_node_id = str(id(child))
+                child_graph_node_id = str(id(child.token if isinstance(child, TokenNode) else child))
                 child_graph_node = self.graph.get_node(child_graph_node_id)[0]
                 self.graph.add_edge(self.pydot.Edge(graph_node, child_graph_node))
             else: