+++ tests/masonry/builders/test_complete.py
@@ -276,12 +276,13 @@ Classifier: Programming Language :: Python :: 3.7
 Classifier: Programming Language :: Python :: 3.8
 Classifier: Programming Language :: Python :: 3.9
 Classifier: Programming Language :: Python :: 3.10
+Classifier: Programming Language :: Python :: 3.11
 Classifier: Topic :: Software Development :: Build Tools
 Classifier: Topic :: Software Development :: Libraries :: Python Modules
 Provides-Extra: time
 Requires-Dist: cachy[msgpack] (>=0.2.0,<0.3.0)
 Requires-Dist: cleo (>=0.6,<0.7)
-Requires-Dist: pendulum (>=1.4,<2.0); (python_version ~= "2.7" and sys_platform == "win32" or python_version in "3.4 3.5") and (extra == "time")
+Requires-Dist: pendulum (>=1.4,<2.0) ; (python_version ~= "2.7" and sys_platform == "win32" or python_version in "3.4 3.5") and (extra == "time")
 Project-URL: Documentation, https://python-poetry.org/docs
 Project-URL: Issue Tracker, https://github.com/python-poetry/poetry/issues
 Project-URL: Repository, https://github.com/python-poetry/poetry
@@ -400,12 +401,13 @@ Classifier: Programming Language :: Python :: 3.7
 Classifier: Programming Language :: Python :: 3.8
 Classifier: Programming Language :: Python :: 3.9
 Classifier: Programming Language :: Python :: 3.10
+Classifier: Programming Language :: Python :: 3.11
 Classifier: Topic :: Software Development :: Build Tools
 Classifier: Topic :: Software Development :: Libraries :: Python Modules
 Provides-Extra: time
 Requires-Dist: cachy[msgpack] (>=0.2.0,<0.3.0)
 Requires-Dist: cleo (>=0.6,<0.7)
-Requires-Dist: pendulum (>=1.4,<2.0); (python_version ~= "2.7" and sys_platform == "win32" or python_version in "3.4 3.5") and (extra == "time")
+Requires-Dist: pendulum (>=1.4,<2.0) ; (python_version ~= "2.7" and sys_platform == "win32" or python_version in "3.4 3.5") and (extra == "time")
 Project-URL: Documentation, https://python-poetry.org/docs
 Project-URL: Issue Tracker, https://github.com/python-poetry/poetry/issues
 Project-URL: Repository, https://github.com/python-poetry/poetry
@@ -425,12 +427,12 @@ def test_module_src() -> None:
     builder = Builder(Factory().create_poetry(module_path))
     builder.build(fmt="all")
 
-    sdist = module_path / "dist" / "module-src-0.1.tar.gz"
+    sdist = module_path / "dist" / "module_src-0.1.tar.gz"
 
     assert sdist.exists()
 
     with tarfile.open(str(sdist), "r") as tar:
-        assert "module-src-0.1/src/module_src.py" in tar.getnames()
+        assert "module_src-0.1/src/module_src.py" in tar.getnames()
 
     whl = module_path / "dist" / "module_src-0.1-py2.py3-none-any.whl"
 
@@ -449,12 +451,12 @@ def test_package_src() -> None:
     builder = Builder(Factory().create_poetry(module_path))
     builder.build(fmt="all")
 
-    sdist = module_path / "dist" / "package-src-0.1.tar.gz"
+    sdist = module_path / "dist" / "package_src-0.1.tar.gz"
 
     assert sdist.exists()
 
     with tarfile.open(str(sdist), "r") as tar:
-        assert "package-src-0.1/src/package_src/module.py" in tar.getnames()
+        assert "package_src-0.1/src/package_src/module.py" in tar.getnames()
 
     whl = module_path / "dist" / "package_src-0.1-py2.py3-none-any.whl"
 
@@ -474,13 +476,13 @@ def test_split_source() -> None:
     builder = Builder(Factory().create_poetry(module_path))
     builder.build(fmt="all")
 
-    sdist = module_path / "dist" / "split-source-0.1.tar.gz"
+    sdist = module_path / "dist" / "split_source-0.1.tar.gz"
 
     assert sdist.exists()
 
     with tarfile.open(str(sdist), "r") as tar:
-        assert "split-source-0.1/lib_a/module_a/__init__.py" in tar.getnames()
-        assert "split-source-0.1/lib_b/module_b/__init__.py" in tar.getnames()
+        assert "split_source-0.1/lib_a/module_a/__init__.py" in tar.getnames()
+        assert "split_source-0.1/lib_b/module_b/__init__.py" in tar.getnames()
 
     whl = module_path / "dist" / "split_source-0.1-py3-none-any.whl"
 
@@ -520,30 +522,30 @@ def test_package_with_include(mocker: MockerFixture) -> None:
     builder = Builder(Factory().create_poetry(module_path))
     builder.build(fmt="all")
 
-    sdist = fixtures_dir / "with-include" / "dist" / "with-include-1.2.3.tar.gz"
+    sdist = fixtures_dir / "with-include" / "dist" / "with_include-1.2.3.tar.gz"
 
     assert sdist.exists()
 
     with tarfile.open(str(sdist), "r") as tar:
         names = tar.getnames()
         assert len(names) == len(set(names))
-        assert "with-include-1.2.3/LICENSE" in names
-        assert "with-include-1.2.3/README.rst" in names
-        assert "with-include-1.2.3/extra_dir/__init__.py" in names
-        assert "with-include-1.2.3/extra_dir/vcs_excluded.txt" in names
-        assert "with-include-1.2.3/extra_dir/sub_pkg/__init__.py" in names
-        assert "with-include-1.2.3/extra_dir/sub_pkg/vcs_excluded.txt" not in names
-        assert "with-include-1.2.3/my_module.py" in names
-        assert "with-include-1.2.3/notes.txt" in names
-        assert "with-include-1.2.3/package_with_include/__init__.py" in names
-        assert "with-include-1.2.3/tests/__init__.py" in names
-        assert "with-include-1.2.3/pyproject.toml" in names
-        assert "with-include-1.2.3/setup.py" in names
-        assert "with-include-1.2.3/PKG-INFO" in names
-        assert "with-include-1.2.3/for_wheel_only/__init__.py" not in names
-        assert "with-include-1.2.3/src/src_package/__init__.py" in names
+        assert "with_include-1.2.3/LICENSE" in names
+        assert "with_include-1.2.3/README.rst" in names
+        assert "with_include-1.2.3/extra_dir/__init__.py" in names
+        assert "with_include-1.2.3/extra_dir/vcs_excluded.txt" in names
+        assert "with_include-1.2.3/extra_dir/sub_pkg/__init__.py" in names
+        assert "with_include-1.2.3/extra_dir/sub_pkg/vcs_excluded.txt" not in names
+        assert "with_include-1.2.3/my_module.py" in names
+        assert "with_include-1.2.3/notes.txt" in names
+        assert "with_include-1.2.3/package_with_include/__init__.py" in names
+        assert "with_include-1.2.3/tests/__init__.py" in names
+        assert "with_include-1.2.3/pyproject.toml" in names
+        assert "with_include-1.2.3/setup.py" in names
+        assert "with_include-1.2.3/PKG-INFO" in names
+        assert "with_include-1.2.3/for_wheel_only/__init__.py" not in names
+        assert "with_include-1.2.3/src/src_package/__init__.py" in names
 
-        file = tar.extractfile("with-include-1.2.3/setup.py")
+        file = tar.extractfile("with_include-1.2.3/setup.py")
         assert file
         setup = file.read()
         setup_ast = ast.parse(setup)
@@ -587,31 +589,31 @@ def test_respect_format_for_explicit_included_files() -> None:
     builder = Builder(Factory().create_poetry(module_path))
     builder.build(fmt="all")
 
-    sdist = module_path / "dist" / "exclude-whl-include-sdist-0.1.0.tar.gz"
+    sdist = module_path / "dist" / "exclude_whl_include_sdist-0.1.0.tar.gz"
 
     assert sdist.exists()
 
     with tarfile.open(str(sdist), "r") as tar:
         names = tar.getnames()
         assert (
-            "exclude-whl-include-sdist-0.1.0/exclude_whl_include_sdist/__init__.py"
+            "exclude_whl_include_sdist-0.1.0/exclude_whl_include_sdist/__init__.py"
             in names
         )
         assert (
-            "exclude-whl-include-sdist-0.1.0/exclude_whl_include_sdist/compiled/source.c"
+            "exclude_whl_include_sdist-0.1.0/exclude_whl_include_sdist/compiled/source.c"
             in names
         )
         assert (
-            "exclude-whl-include-sdist-0.1.0/exclude_whl_include_sdist/compiled/source.h"
+            "exclude_whl_include_sdist-0.1.0/exclude_whl_include_sdist/compiled/source.h"
             in names
         )
         assert (
-            "exclude-whl-include-sdist-0.1.0/exclude_whl_include_sdist/cython_code.pyx"
+            "exclude_whl_include_sdist-0.1.0/exclude_whl_include_sdist/cython_code.pyx"
             in names
         )
-        assert "exclude-whl-include-sdist-0.1.0/pyproject.toml" in names
-        assert "exclude-whl-include-sdist-0.1.0/setup.py" in names
-        assert "exclude-whl-include-sdist-0.1.0/PKG-INFO" in names
+        assert "exclude_whl_include_sdist-0.1.0/pyproject.toml" in names
+        assert "exclude_whl_include_sdist-0.1.0/setup.py" in names
+        assert "exclude_whl_include_sdist-0.1.0/PKG-INFO" in names
 
     whl = module_path / "dist" / "exclude_whl_include_sdist-0.1.0-py3-none-any.whl"