+++ tests/packages/test_directory_dependency.py
@@ -26,7 +26,7 @@ def _test_directory_dependency_pep_508(
     )
 
     assert dep.is_directory()
-    dep = cast(DirectoryDependency, dep)
+    dep = cast("DirectoryDependency", dep)
     assert dep.name == name
     assert dep.path == path
     assert dep.to_pep_508() == (pep_508_output or pep_508_input)
@@ -66,7 +66,8 @@ def test_directory_dependency_pep_508_local_relative() -> None:
         _test_directory_dependency_pep_508("demo", path, requirement)
 
     requirement = f"demo @ {path}"
-    expected = f"demo @ {path.as_posix()}"
+    base = Path(__file__).parent
+    expected = f"demo @ {(base / path).resolve().as_uri()}"
     _test_directory_dependency_pep_508("demo", path, requirement, expected)
 
 
@@ -78,6 +79,17 @@ def test_directory_dependency_pep_508_extras() -> None:
     )
     requirement = f"demo[foo,bar] @ file://{path.as_posix()}"
     expected = f"demo[bar,foo] @ {path.as_uri()}"
+    _test_directory_dependency_pep_508("demo", path, requirement, expected)
+
+
+def test_directory_dependency_pep_508_with_marker() -> None:
+    path = (
+        Path(__file__).parent.parent
+        / "fixtures"
+        / "project_with_multi_constraints_dependency"
+    )
+    requirement = f'demo @ file://{path.as_posix()} ; sys_platform == "linux"'
+    expected = f'demo @ {path.as_uri()} ; sys_platform == "linux"'
     _test_directory_dependency_pep_508("demo", path, requirement, expected)